Fix build warnings

+++ src/C++/HttpConnection.cpp
@@ -61,7 +61,7 @@ bool HttpConnection::read()
   try
   {
     // Wait for input (1 second timeout)
-    int result = select( 1 + m_socket, &readset, 0, 0, &timeout );
+    int result = select( 1 + (int)m_socket, &readset, 0, 0, &timeout );
 
     if( result > 0 ) // Something to read
     {
