+++ src/C++/ThreadedSSLSocketConnection.cpp
@@ -227,7 +227,7 @@ bool ThreadedSSLSocketConnection::read()
   try
   {
     // Wait for input (1 second timeout)
-    int result = select(1 + m_socket, &readset, 0, 0, &timeout);
+    int result = select(1 + (int)m_socket, &readset, 0, 0, &timeout);
 
     if (result > 0) // Something to read
     {
